requirements.txt
PyYAML==3.11
neo4j-driver==1.0.2
